Put views into different files and no magic numbers
